requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0

# LLM API Clients
openai>=1.0.0
anthropic>=0.18.0
google-genai>=1.0.0
